chore(earn): move earn report related code to own dir
