+++ src/components/earn/report/EarnReportSheet.tsx
@@ -13,6 +13,8 @@ import {
 import { DownloadIcon, PlusIcon, RefreshCwIcon, SearchIcon } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
 import { useStore } from 'zustand'
+import { DevBadge } from '@/components/dev/DevBadge'
+import { useQueryYieldgenReport, type EarnReportEntry } from '@/components/earn/report/hooks/useQueryYieldgenReport'
 import { Alert, AlertTitle } from '@/components/ui/alert'
 import { Button } from '@/components/ui/button'
 import { Card, CardContent } from '@/components/ui/card'
@@ -23,11 +25,9 @@ import { SortIcon } from '@/components/ui/jam/SortIcon'
 import { TablePagination } from '@/components/ui/jam/TablePagination'
 import { Spinner } from '@/components/ui/spinner'
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table'
-import { useQueryYieldgenReport, type EarnReportEntry } from '@/hooks/useQueryYieldgenReport'
 import { pseudoRandomFloat, pseudoRandomInteger } from '@/lib/utils'
 import { jamSettingsStore } from '@/store/jamSettingsStore'
 import type { AmountSats, Milliseconds } from '@/types/global'
-import { DevBadge } from '../dev/DevBadge'
 import { EarnReportChart } from './EarnReportChart'
 
 // Bitcoin genesis block date - used as the default 'since' for all-time sums
